+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- MS3Record.samprate() consistently returns the rate in samples/second,
-  and MS3Record.samplrate_raw() returns the record's raw value.
+- MS3Record.samprate consistently returns the rate in samples/second,
+  and MS3Record.samprate_raw returns the record's raw value.
 
+### Added
+- MS3Record.samprate_period_ns to return sample period in nanoseconds.
+- MS3Record.samprate_period_seconds to return the period in seconds.
 
 ## [0.0.4] - 2025-9-24
 ### Fixed